requirement_local.txt
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.11.5
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
crispy-bootstrap5==0.6
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.4
Django==4.0
django-allauth==0.50.0
django-cors-headers==3.13.0
django-coverage-plugin==2.0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==22.1
django-model-utils==4.2.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-stubs==1.11.0
django-stubs-ext==0.4.0
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-stubs==1.6.0
docutils==0.18.1
drf-spectacular==0.22.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.12.1
filelock==3.7.1
flake8==4.0.1
flake8-isort==4.1.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-HTTPAuth==4.7.0
Flask-SQLAlchemy==2.4.0
hiredis==2.0.0
identify==2.5.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.0
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.14.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
sandman2==1.2.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.3
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tornado==6.1
traitlets==5.2.2.post1
types-Markdown==3.3.27
types-pytz==2021.3.8
types-PyYAML==6.0.8
types-requests==2.27.30
types-urllib3==1.26.15
typing-extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
watchdog==2.1.8
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==6.2.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
